refactor: dedupe reportStatus/expenseType key maps into packages/i18n
apps/web/src/lib/i18n-labels.ts and apps/api/src/services/pdf/format.ts
each defined an identical enum-value-to-translation-key map, risking
drift whenever a status or expense type is added. Move them into a new
packages/i18n/src/enums.ts, typed against locally-declared string
literal unions rather than importing @zemio/db's Prisma-generated
types, so the i18n package stays independent of the ORM package. Both
apps now import the single shared map.
