+++ packages/i18n/src/index.ts
@@ -1,4 +1,10 @@
 export {
+	type ExpenseTypeValue,
+	expenseTypeKeys,
+	type ReportStatusValue,
+	reportStatusKeys,
+} from "./enums";
+export {
 	defaultLocale,
 	getMessages,
 	type Locale,
